Create singlesignallearn.py
This is the single signal training, ECOS solver fails very often. We may need to download MOSDEK, GUROBI, CVXOPT and use in place of ECOS. You may check if I used the opti.zero() or the requires grad at the correct places. Let me know if it fails to work. The main problem is still with the ECOS solver.