Create  Tv_closedform
This code proves that closed form expression works when the correct signs are obtained. An important observation: the threshold and lamd is very sensitive. lamd can be taken care of by exhaustive search. But the threshold is to be tuned carefully. This maybe one of the reasons why gradient descent is not converging.